@@ -18,13 +18,13 @@
     "lint:strict": "eslint \"**/*.{js,jsx,ts,tsx}\" --max-warnings 0",
     "lint:types": "tsc --noEmit",
     "lint:all": "yarn lint && yarn lint:types && yarn lint:strict",
+    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
+    "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "analyze": "ANALYZE=true next build",
     "migrate:all": "npx prisma migrate deploy",
     "validate:env": "ts-node scripts/validate-env.ts",
     "backup": "ts-node scripts/backup.ts",
     "backup:schedule": "node -r ts-node/register scripts/schedule-backup.ts",
-    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
-    "format:check": "prettier --check .",
     "prepare": "husky"
   },
   "prisma": {
@@ -64,7 +64,7 @@
     "@svgr/webpack": "^8.1.0",
     "@tanstack/react-query": "^5.69.0",
     "@tanstack/react-table": "^8.21.3",
-    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/eslint-plugin": "6.21.0",
     "@upstash/ratelimit": "^2.0.5",
     "@upstash/redis": "^1.35.0",
     "@vercel/analytics": "^1.5.0",
@@ -117,10 +117,12 @@
   },
   "devDependencies": {
     "@eslint/js": "^8.56.0",
+    "@faker-js/faker": "^9.8.0",
     "@playwright/test": "^1.52.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
+    "@types/bcryptjs": "^3.0.0",
     "@types/cors": "^2.8.17",
     "@types/diff": "^8.0.0",
     "@types/express": "^4.17.21",
@@ -131,16 +133,15 @@
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/parser": "^6.21.0",
+    "@typescript-eslint/parser": "6.21.0",
     "@vitejs/plugin-react": "^4.5.1",
     "autoprefixer": "10.4.16",
+    "bcryptjs": "^3.0.2",
     "eslint": "^8.56.0",
     "eslint-config-next": "14.1.0",
-    "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "globals": "^14.0.0",
@@ -149,12 +150,12 @@
     "lint-staged": "^16.1.0",
     "node-mocks-http": "^1.17.2",
     "postcss": "8.4.31",
-    "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.11",
+    "prettier": "^3.5.3",
+    "prettier-plugin-tailwindcss": "^0.6.12",
     "tailwindcss": "3.3.0",
     "ts-jest": "^29.3.2",
     "ts-node": "^10.9.2",
-    "typescript": "^5",
+    "typescript": "^5.3.3",
     "typescript-eslint": "^8.33.1",
     "vitest": "^3.2.2"
   }
