@@ -2,13 +2,23 @@
 
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import { Users, BarChart2, Settings, Shield, LayoutDashboard } from 'lucide-react';
+import {
+  Users,
+  BarChart2,
+  Settings,
+  Shield,
+  LayoutDashboard,
+  MessageSquare,
+  FileText,
+} from 'lucide-react';
 
 const navigation = [
   { name: 'Dashboard', href: '/admin', icon: LayoutDashboard },
   { name: 'Users', href: '/admin/users', icon: Users },
   { name: 'Analytics', href: '/admin/analytics', icon: BarChart2 },
   { name: 'Moderation', href: '/admin/moderation', icon: Shield },
+  { name: 'Support Tickets', href: '/admin/support', icon: MessageSquare },
+  { name: 'Email Templates', href: '/admin/email-templates', icon: FileText },
   { name: 'Settings', href: '/admin/settings', icon: Settings },
 ];
 
@@ -16,11 +26,11 @@ export default function AdminSidebar() {
   const pathname = usePathname();
 
   return (
-    <div className="w-64 border-r border-gray-200 bg-white">
+    <div className="flex h-full flex-col border-r border-gray-200 bg-white">
       <div className="flex h-16 items-center border-b border-gray-200 px-6">
         <h1 className="text-xl font-semibold text-gray-900">Admin Panel</h1>
       </div>
-      <nav className="mt-6">
+      <nav className="flex-1 overflow-y-auto py-4">
         {navigation.map(item => {
           const isActive = pathname === item.href;
           return (
