@@ -3,6 +3,7 @@ import { auth } from '@clerk/nextjs/server';
 import { SupportService } from '@/lib/services/supportService';
 import { z } from 'zod';
 import { dynamicRouteConfig, withDynamicRoute } from '@/lib/utils/dynamicRoute';
+import { TicketStatus } from '@prisma/client';
 
 // Export dynamic configuration
 export const { dynamic, revalidate, runtime } = dynamicRouteConfig;
@@ -27,7 +28,7 @@ async function ticketDetailHandler(
   }
 
   const supportService = SupportService.getInstance();
-  const ticket = await supportService.getTicket(ticketId, userId);
+  const ticket = await supportService.getTicket(ticketId);
 
   if (!ticket) {
     return NextResponse.json({ error: 'Ticket not found' }, { status: 404 });
@@ -37,10 +38,7 @@ async function ticketDetailHandler(
 }
 
 // Define the POST handler
-async function addReplyHandler(
-  request: Request,
-  context?: { params?: Record<string, string> }
-) {
+async function addReplyHandler(request: Request, context?: { params?: Record<string, string> }) {
   try {
     const ticketId = context?.params?.id;
     if (!ticketId) {
@@ -79,7 +77,7 @@ export const POST = withDynamicRoute(addReplyHandler, fallbackData);
 
 export async function PATCH(
   request: Request,
-  // @ts-ignore
+  // @ts-expect-error params is typed as any
   { params }
 ) {
   try {
@@ -95,8 +93,13 @@ export async function PATCH(
       return NextResponse.json({ error: 'Status is required' }, { status: 400 });
     }
 
+    const validStatuses: readonly TicketStatus[] = ['OPEN', 'IN_PROGRESS', 'RESOLVED', 'CLOSED'];
+    if (!validStatuses.includes(status as TicketStatus)) {
+      return NextResponse.json({ error: 'Invalid status' }, { status: 400 });
+    }
+    const ticketStatus = status as TicketStatus;
     const supportService = SupportService.getInstance();
-    const ticket = await supportService.updateTicketStatus(params.id, userId, status);
+    const ticket = await supportService.updateTicketStatus(params.id, ticketStatus, userId);
 
     return NextResponse.json(ticket);
   } catch (error) {
