@@ -3,6 +3,8 @@ import { auth } from '@clerk/nextjs/server';
 import { SupportService } from '@/lib/services/supportService';
 import { z } from 'zod';
 import { dynamicRouteConfig, withDynamicRoute } from '@/lib/utils/dynamicRoute';
+import { Category, Priority, TicketStatus } from '@prisma/client';
+import { prisma } from '@/lib/prisma';
 
 // Export dynamic configuration
 export const { dynamic, revalidate, runtime } = dynamicRouteConfig;
@@ -10,53 +12,139 @@ export const { dynamic, revalidate, runtime } = dynamicRouteConfig;
 const ticketSchema = z.object({
   title: z.string().min(1, 'Title is required'),
   description: z.string().min(1, 'Description is required'),
-  category: z.enum(['BUG', 'FEATURE_REQUEST', 'GENERAL_INQUIRY', 'TECHNICAL_SUPPORT', 'BILLING']),
-  priority: z.enum(['LOW', 'MEDIUM', 'HIGH', 'URGENT']).optional(),
+  category: z.string().min(1, 'Category is required'),
+  priority: z.string().optional(),
 });
+
+// Add CORS headers helper
+function corsHeaders() {
+  return {
+    'Access-Control-Allow-Origin': '*',
+    'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE, OPTIONS',
+    'Access-Control-Allow-Headers': 'Content-Type, Authorization',
+  };
+}
+
+// Handle OPTIONS request for CORS preflight
+export async function OPTIONS() {
+  return NextResponse.json({}, { headers: corsHeaders() });
+}
 
 // Define the main handler
 async function ticketsHandler(request: Request) {
-  const { userId } = await auth();
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  try {
+    const { userId: clerkId } = await auth();
+    if (!clerkId) {
+      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
+    }
+
+    const { searchParams } = new URL(request.url);
+    const status = searchParams.get('status');
+    const category = searchParams.get('category');
+    const priority = searchParams.get('priority');
+
+    // Create filters object only with valid values
+    const filters: {
+      status?: TicketStatus;
+      category?: Category;
+      priority?: Priority;
+    } = {};
+
+    if (status && Object.values(TicketStatus).includes(status as TicketStatus)) {
+      filters.status = status as TicketStatus;
+    }
+    if (category && Object.values(Category).includes(category as Category)) {
+      filters.category = category as Category;
+    }
+    if (priority && Object.values(Priority).includes(priority as Priority)) {
+      filters.priority = priority as Priority;
+    }
+
+    const supportService = SupportService.getInstance();
+    const tickets = await supportService.getTickets(clerkId, filters);
+
+    return NextResponse.json(tickets);
+  } catch (error) {
+    console.error('Error fetching tickets:', error);
+    if (error instanceof Error && error.message === 'User not found in database') {
+      return NextResponse.json(
+        { error: 'User account not found. Please contact support.' },
+        { status: 404 }
+      );
+    }
+    return NextResponse.json({ error: 'Failed to fetch tickets' }, { status: 500 });
   }
-
-  const { searchParams } = new URL(request.url);
-  const status = searchParams.get('status');
-  const category = searchParams.get('category');
-  const priority = searchParams.get('priority');
-
-  const supportService = SupportService.getInstance();
-  const tickets = await supportService.getTickets(userId, {
-    status: status as any,
-    category: category as any,
-    priority: priority as any,
-  });
-
-  return NextResponse.json(tickets);
 }
 
 // Define the POST handler
-async function createTicketHandler(request: Request) {
+export async function POST(req: Request) {
+  console.log('API: Support ticket creation request received');
+
   try {
-    const { userId } = await auth();
+    const session = await auth();
+    const userId = session.userId;
+    console.log('API: Auth check - User ID:', userId);
+
     if (!userId) {
+      console.error('API: Authentication failed - No user ID');
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
 
-    const body = await request.json();
-    const validatedData = ticketSchema.parse(body);
+    const body = await req.json();
+    console.log('API: Request body received:', { ...body, userId: undefined });
 
-    const supportService = SupportService.getInstance();
-    const ticket = await supportService.createTicket(userId, validatedData);
+    // Validate request body
+    const validationResult = ticketSchema.safeParse(body);
+    if (!validationResult.success) {
+      console.error('API: Validation failed:', validationResult.error);
+      return NextResponse.json(
+        { error: 'Invalid request data', details: validationResult.error.format() },
+        { status: 400 }
+      );
+    }
 
+    // Get DB user ID
+    const dbUser = await prisma.user.findUnique({
+      where: { clerkId: userId },
+      select: { id: true },
+    });
+
+    if (!dbUser) {
+      console.error('API: User not found in database:', userId);
+      return NextResponse.json({ error: 'User not found' }, { status: 404 });
+    }
+
+    console.log('API: Creating ticket for user:', dbUser.id);
+
+    // Create ticket
+    const ticket = await prisma.supportTicket.create({
+      data: {
+        title: body.title,
+        description: body.description,
+        category: body.category,
+        priority: body.priority || 'MEDIUM',
+        userId: dbUser.id,
+        status: 'OPEN',
+      },
+      include: {
+        user: {
+          select: {
+            id: true,
+            name: true,
+            imageUrl: true,
+          },
+        },
+      },
+    });
+
+    console.log('API: Ticket created successfully:', { ticketId: ticket.id });
     return NextResponse.json(ticket);
   } catch (error) {
-    if (error instanceof z.ZodError) {
-      return NextResponse.json({ error: error.errors }, { status: 400 });
-    }
-    console.error('Error creating ticket:', error);
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
+    console.error(
+      'API: Error creating ticket:',
+      error instanceof Error ? error.message : 'Unknown error'
+    );
+    return NextResponse.json({ error: 'Failed to create ticket' }, { status: 500 });
   }
 }
 
@@ -67,4 +155,3 @@ const fallbackData = {
 
 // Export the wrapped handlers
 export const GET = withDynamicRoute(ticketsHandler, fallbackData);
-export const POST = withDynamicRoute(createTicketHandler, fallbackData);
