@@ -21,7 +21,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en">
       <body className={inter.className} suppressHydrationWarning>
-        <ErrorBoundary>
+        <ErrorBoundary fallback={<div>Error</div>}>
           <ClerkProvider>
             <Providers>
               <ThemeProvider>
