@@ -65,8 +65,8 @@ export default function Providers({ children }: { children: React.ReactNode }) {
 
   if (!isClient || isLoading) {
     return (
-      <div className="min-h-screen flex items-center justify-center bg-gray-50 dark:bg-gray-900">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-500"></div>
+      <div className="flex min-h-screen items-center justify-center bg-gray-50 dark:bg-gray-900">
+        <div className="h-8 w-8 animate-spin rounded-full border-b-2 border-blue-500"></div>
       </div>
     );
   }
@@ -76,9 +76,9 @@ export default function Providers({ children }: { children: React.ReactNode }) {
       onError={handleError}
       resetOnPropsChange={true}
       fallback={
-        <div className="min-h-screen flex items-center justify-center bg-gray-50 dark:bg-gray-900">
+        <div className="flex min-h-screen items-center justify-center bg-gray-50 dark:bg-gray-900">
           <div className="text-center">
-            <h1 className="text-2xl font-bold text-gray-900 dark:text-white mb-4">
+            <h1 className="mb-4 text-2xl font-bold text-gray-900 dark:text-white">
               Application Error
             </h1>
             <p className="text-gray-600 dark:text-gray-300">
