@@ -27,9 +27,9 @@ export class BillingService {
     const subscription = subscriptions.data[0] || null;
 
     // Fetch invoices
-    const invoices = await stripe.invoices.list({ 
-      customer: user.stripeCustomerId, 
-      limit: 10 
+    const invoices = await stripe.invoices.list({
+      customer: user.stripeCustomerId,
+      limit: 10,
     });
 
     // Fetch payment methods
@@ -61,4 +61,4 @@ export class BillingService {
 
     return portalSession.url;
   }
-} 
+}
