@@ -1,49 +1,30 @@
-import { LRUCache } from 'lru-cache';
+import { Redis } from '@upstash/redis';
+import { Ratelimit } from '@upstash/ratelimit';
+import { NextRequest, NextResponse } from 'next/server';
 
-export interface RateLimitOptions {
-  interval: number;
-  uniqueTokenPerInterval: number;
+const redis = Redis.fromEnv();
+
+export const rateLimiter = new Ratelimit({
+  redis,
+  limiter: Ratelimit.slidingWindow(10, '10 s'),
+  analytics: true,
+  prefix: 'ratelimit',
+});
+
+export async function rateLimit(req: NextRequest): Promise<NextResponse | null> {
+  const ip = req.ip ?? '127.0.0.1';
+  const { success, limit, reset, remaining } = await rateLimiter.limit(ip);
+
+  if (!success) {
+    return new NextResponse('Too Many Requests', {
+      status: 429,
+      headers: {
+        'X-RateLimit-Limit': limit.toString(),
+        'X-RateLimit-Remaining': remaining.toString(),
+        'X-RateLimit-Reset': reset.toString(),
+      },
+    });
+  }
+
+  return null;
 }
-
-export interface RateLimitResult {
-  success: boolean;
-  limit: number;
-  remaining: number;
-  reset: number;
-}
-
-export function rateLimit(options: RateLimitOptions) {
-  const tokenCache = new LRUCache({
-    max: options.uniqueTokenPerInterval || 500,
-    ttl: options.interval || 60000,
-  });
-
-  return {
-    check: (limit: number, token: string = 'default'): Promise<RateLimitResult> => {
-      const tokenCount = (tokenCache.get(token) as number[]) || [0];
-      const now = Date.now();
-      const windowStart = now - options.interval;
-
-      // Remove old requests
-      while (tokenCount.length && tokenCount[0] < windowStart) {
-        tokenCount.shift();
-      }
-
-      // Check if limit is exceeded
-      if (tokenCount.length >= limit) {
-        return Promise.reject(new Error('Rate limit exceeded'));
-      }
-
-      // Add current request
-      tokenCount.push(now);
-      tokenCache.set(token, tokenCount);
-
-      return Promise.resolve({
-        success: true,
-        limit,
-        remaining: limit - tokenCount.length,
-        reset: windowStart + options.interval,
-      });
-    },
-  };
-} 
