@@ -1,5 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { QuotaService } from '@/services/quota-service';
+import { QuotaService } from '@/lib/services/quota-service';
+
+type QuotaType = 'API_CALLS' | 'PROMPT_GENERATIONS' | 'STORAGE';
 
 export async function quotaMiddleware(req: NextRequest) {
   const quotaService = QuotaService.getInstance();
@@ -10,7 +12,7 @@ export async function quotaMiddleware(req: NextRequest) {
   }
 
   // Determine quota type based on the request path
-  let quotaType: 'API_CALLS' | 'PROMPT_GENERATIONS' | 'STORAGE';
+  let quotaType: QuotaType;
   if (req.nextUrl.pathname.startsWith('/api/prompts/generate')) {
     quotaType = 'PROMPT_GENERATIONS';
   } else if (req.nextUrl.pathname.startsWith('/api/storage')) {
@@ -22,21 +24,21 @@ export async function quotaMiddleware(req: NextRequest) {
   // Check quota
   const quotaResult = await quotaService.checkQuota(userId, quotaType);
 
-  if (!quotaResult.isAllowed) {
+  if (!quotaResult.allowed) {
     return new NextResponse(
       JSON.stringify({
         error: 'Quota exceeded',
         message: `You have exceeded your ${quotaType.toLowerCase()} quota. Please upgrade your plan or try again later.`,
-        remaining: quotaResult.remaining,
-        limit: quotaResult.limit,
-        resetAt: quotaResult.resetAt,
+        remaining: quotaResult.remaining ?? 0,
+        total: quotaResult.total ?? 0,
+        resetAt: quotaResult.resetAt ?? null,
       }),
       {
         status: 429,
         headers: {
           'Content-Type': 'application/json',
-          'X-RateLimit-Limit': quotaResult.limit.toString(),
-          'X-RateLimit-Remaining': quotaResult.remaining.toString(),
+          'X-RateLimit-Limit': (quotaResult.total ?? 0).toString(),
+          'X-RateLimit-Remaining': (quotaResult.remaining ?? 0).toString(),
           'X-RateLimit-Reset': quotaResult.resetAt?.toISOString() || '',
         },
       }
@@ -45,8 +47,8 @@ export async function quotaMiddleware(req: NextRequest) {
 
   // Add quota headers to response
   const response = NextResponse.next();
-  response.headers.set('X-RateLimit-Limit', quotaResult.limit.toString());
-  response.headers.set('X-RateLimit-Remaining', quotaResult.remaining.toString());
+  response.headers.set('X-RateLimit-Limit', (quotaResult.total ?? 0).toString());
+  response.headers.set('X-RateLimit-Remaining', (quotaResult.remaining ?? 0).toString());
   response.headers.set('X-RateLimit-Reset', quotaResult.resetAt?.toISOString() || '');
 
   return response;
