@@ -17,7 +17,7 @@ vi.mock('next/navigation', () => ({
 // Mock next/image
 vi.mock('next/image', () => ({
   default: (props: any) => {
-    // eslint-disable-next-line @next/next/no-img-element
+    // eslint-disable-next-line react/no-danger
     return React.createElement('img', props);
   },
 }));
